Renaming_logdetailsdto
