+++ src/app/Models/logdetails-dto.ts
